requirements.txt
hupper==1.1
Mako==1.0.7
MarkupSafe==1.0
PasteDeploy==1.5.2
plaster==1.0
plaster-pastedeploy==0.5
Pygments==2.2.0
pymongo==3.6.1
python-dateutil==2.7.2
pyramid==1.9.1
pyramid-debugtoolbar==4.4
pyramid-mako==1.0.2
repoze.lru==0.7
translationstring==1.3
validate-email==1.3
venusian==1.1.0
waitress==1.1.0
WebOb==1.8.1
wsgicors==0.7.0
zope.deprecation==4.3.0
zope.interface==4.4.3
